UI_&_Events
-Se creó el menú principal en una nueva scene, se coloco en un canvas un background, un texto y un botón para iniciar el juego.
-Se creo un evento donde al perder vida se llama al HUD (canvas) y nos quita un corazón de vida.

+++ Platformer_Unity_PV2/Assets/Scripts/PerfilJugador.cs
@@ -27,8 +27,8 @@ public class PerfilJugador : ScriptableObject
 
 
     [Header("Lifes")]
-    [SerializeField] private float vida = 5f;
-    public float Vida { get => vida; set => vida = value; }
+    [SerializeField] private int vida = 5;
+    public int Vida { get => vida; set => vida = value; }
     [SerializeField] private float InmortalTime;
     public float InmortalTime1 { get => InmortalTime; set => InmortalTime = value; }
     [SerializeField] private  float intervaloParpadeo = 0.5f;
